Conflicts:
	src/main/java/nl/focalor/utobot/hipchat/model/HipchatSettings.java
	src/main/resources/hipchat-settings.json
	src/main/resources/logback.xml

+++ src/main/java/nl/focalor/utobot/utopia/service/ProvinceService.java
@@ -1,14 +1,18 @@
 package nl.focalor.utobot.utopia.service;
 
+import nl.focalor.utobot.base.model.entity.Person;
+import nl.focalor.utobot.base.model.service.IPersonService;
 import nl.focalor.utobot.utopia.model.entity.Province;
 import nl.focalor.utobot.utopia.model.repository.ProvinceRepository;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class ProvinceService implements IProvinceService {
-
+	@Autowired
+	private IPersonService personService;
 	@Autowired
 	private ProvinceRepository provinceDao;
 
@@ -27,7 +31,14 @@ public class ProvinceService implements IProvinceService {
 	@Override
 	@Transactional
 	public void deleteByNameIgnoreCase(String name) {
-		provinceDao.deleteByNameIgnoreCase(name);
+		Province prov = provinceDao.findByNameIgnoreCase(name);
+		if (prov != null) {
+			Person owner = prov.getOwner();
+			owner.setProvince(null);
+			personService.save(owner);
+
+			provinceDao.deleteByNameIgnoreCase(name);
+		}
 	}
 
 	@Override
